+++ config/workdays-iran.php
@@ -25,6 +25,8 @@ return [
 
     'include_start_date' => false,
 
+    'max_scan_days' => 3660,
+
     'storage' => [
         'driver' => 'config',
     ],