+++ tests/TestCase.php
@@ -4,13 +4,14 @@ declare(strict_types=1);
 
 namespace Zarbinco\LaravelWorkdays\Tests;
 
+use Illuminate\Foundation\Application;
 use Orchestra\Testbench\TestCase as Orchestra;
 use Zarbinco\LaravelWorkdays\WorkdaysServiceProvider;
 
 abstract class TestCase extends Orchestra
 {
     /**
-     * @param \Illuminate\Foundation\Application $app
+     * @param  Application  $app
      * @return array<int, class-string>
      */
     protected function getPackageProviders($app): array
@@ -21,7 +22,7 @@ abstract class TestCase extends Orchestra
     }
 
     /**
-     * @param \Illuminate\Foundation\Application $app
+     * @param  Application  $app
      */
     protected function defineEnvironment($app): void
     {